Move PropertyValidationAccess over to :plugin-development
